Refactor code for improved readability and maintainability
- Simplified the logic for finding duplicate keys in `find_duplicate_keys.dart`.
- Updated `lifecycle_engine_test.dart` to use `const` for `MaterialApp` instantiation.
- Removed outdated TODO comments in various files, including `board_top_slider.dart`, `custom_drawer_item.dart`, and `replay_service.dart`.
- Cleaned up imports and restructured code in `mill.dart` and related files for better organization.
- Enhanced animation methods in `animation_manager.dart` to ensure proper disposal checks.
- Adjusted game controller logic in `game_controller.dart` for clarity and consistency.
- Updated comments and documentation across multiple files to reflect current functionality and future improvements.

+++ src/ui/flutter_app/lib/games/mill/native_mill_ai_turn_controller.dart
@@ -75,9 +75,9 @@ class NativeMillAiTurnController {
 
   int searchDepthForSnapshot(GameStateSnapshot snapshot) {
     if (depth != null) {
-      return depth!.clamp(1, 64).toInt();
+      return depth!.clamp(1, 64);
     }
-    final int level = generalSettings.skillLevel.clamp(1, 30).toInt();
+    final int level = generalSettings.skillLevel.clamp(1, 30);
     if (!generalSettings.drawOnHumanExperience || snapshot.phase != 'placing') {
       return level;
     }
@@ -94,11 +94,8 @@ class NativeMillAiTurnController {
     final int whiteTotal = whiteInHand + whiteOnBoard;
     final int blackTotal = blackInHand + blackOnBoard;
     final int pieceCount = (whiteTotal > blackTotal ? whiteTotal : blackTotal)
-        .clamp(0, 12)
-        .toInt();
-    final int index = (pieceCount * 2 - whiteInHand - blackInHand)
-        .clamp(0, 24)
-        .toInt();
+        .clamp(0, 12);
+    final int index = (pieceCount * 2 - whiteInHand - blackInHand).clamp(0, 24);
 
     final List<int> table = pieceCount == 12
         ? _placingDepthTable12
